+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### 💫 New checkers
 
-- `unmentioned-parameter`: Provide error message when a function parameter is not mentioned by name in the function's docstring.
+- `unmentioned-parameter`: Provide error message when a function parameter is not mentioned by name in the function's docstring. By default, this checker is disabled.
 
 ### 🐛 Bug fixes